requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
Pillow>=10.0.0
tqdm>=4.66.0
Flask>=3.0.0
python-dotenv>=1.0.0
croniter>=2.0.0
croniter
